feat(frog-crossing): Add datageneration and visualization of frog-crossing game (#3)
I was not able to get the code to reliably cross the grid when the cars
were moving.

Because the two asynchronous systems (playwright test runner,
FrogCrossing component) each with their own timing.
In order for the test to reliably predict movement of cars and avoid
would likely take a much more sophisticated algorithm.

I reduced the complexity by changing the game to only crossing a static
configuration of cars.

## Visualizer Samples


![image](https://github.com/user-attachments/assets/f052b678-72f7-4eca-a7f0-b18e49cf2b0e)


![image](https://github.com/user-attachments/assets/f5d50160-e813-4176-bd0b-d2443466e242)

I was not able to upload data